Refactor para programación nueva

+++ source_code/include/leds.h
@@ -12,7 +12,7 @@
 
 void set_status_led(bool state);
 void toggle_status_led(void);
-void warning_status_led(uint16_t ms);
+void warning_status_led(uint32_t ms);
 void set_RGB_color(uint32_t r, uint32_t g, uint32_t b);
 void set_RGB_rainbow(void);
 void set_leds_wave(int ms);
